+ start of refactoring
