+++ .idea/dictionaries/Dictionary.xml
@@ -1,0 +1,9 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="Dictionary">
+    <words>
+      <w>omdb</w>
+      <w>stapi</w>
+      <w>tmdb</w>
+    </words>
+  </dictionary>
+</component>
