+++ app/src/main/java/com/sevenzeroes/trekkieapp/list/domain/MovieApi.kt
@@ -12,8 +12,6 @@ interface MovieApi {
     }
 
     @GET("3/tv/655/season/{season_number}/episode/{episode_number}?api_key=$API_KEY&language=en-US")
-    suspend fun getEpisodeSpecifics(
-        @Path("season_number") season_number: Int, @Path("episode_number") episode_number : Int
-    ): Response<JsonElement>
+    suspend fun getEpisodeSpecifics(@Path("season_number") season_number: Int, @Path("episode_number") episode_number : Int): Response<JsonElement>
 
 }
