+++ app/src/main/java/com/sevenzeroes/trekkieapp/list/ui/EpisodesFragment.kt
@@ -6,16 +6,11 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.lifecycle.lifecycleScope
 import com.sevenzeroes.trekkieapp.databinding.EpisodesFragmentBinding
-import com.sevenzeroes.trekkieapp.list.data.EpisodeSummary
-import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 
 class EpisodesFragment : Fragment() {
-
-    companion object {
-        fun newInstance() = EpisodesFragment()
-    }
 
     private lateinit var episodesViewModel: EpisodesViewModel
     private lateinit var binding: EpisodesFragmentBinding
@@ -33,8 +28,8 @@ class EpisodesFragment : Fragment() {
 
         setupEpisodesList()
         observeEpisodes()
-        GlobalScope.launch {
-        episodesViewModel.getEpisodesWithLiveData("Realm of")
+        viewLifecycleOwner.lifecycleScope.launch {
+        episodesViewModel.getEpisodes("realm of")
         }
     }
 
@@ -52,125 +47,4 @@ class EpisodesFragment : Fragment() {
             adapter.setData(it)
         })
     }
-/*
-    class MainActivity : AppCompatActivity()  {
-
-        private lateinit var activityViewModel: ActivityViewModel
-        private lateinit var binding: ActivityMainBinding
-        private lateinit var contentMainBinding: ContentMainBinding
-        private lateinit var rvEpisodeList: RecyclerView
-        private lateinit var slEpisodeList: ShimmerFrameLayout
-        private lateinit var tvPrompt: TextView
-        private lateinit var toolbar: androidx.appcompat.widget.Toolbar
-
-
-        override fun onCreate(savedInstanceState: Bundle?) {
-            super.onCreate(savedInstanceState)
-            activityViewModel = ViewModelProviders.of(this).get(ActivityViewModel::class.java)
-            binding = ActivityMainBinding.inflate(layoutInflater)
-            contentMainBinding = ContentMainBinding.inflate(layoutInflater)
-            val view = binding.root
-            setContentView(view)
-            configureViews()
-            observeIsLoading()
-            observeGetPosts()
-        }
-
-        private fun configureViews(){
-            rvEpisodeList = contentMainBinding.rvEpisodeList
-            slEpisodeList = contentMainBinding.slEpisodeList
-            tvPrompt = contentMainBinding.tvPrompt
-            toolbar = binding.toolbar
-            setSupportActionBar(toolbar)
-            supportActionBar?.setDisplayShowTitleEnabled(false)
-            slEpisodeList.hide()
-            rvEpisodeList.show()
-            tvPrompt.show()
-            rvEpisodeList.addItemDecoration(DividerItemDecoration(this, LinearLayoutManager.VERTICAL)) //good stuff
-
-        }
-
-
-        private fun observeIsLoading() {
-            activityViewModel.isLoading.observe(this, Observer {
-                if (it){
-                    slEpisodeList.show()
-                    slEpisodeList.startShimmer()
-                    rvEpisodeList.hide()
-                    togglePromptVisibility(false)
-                }
-                else {
-                    slEpisodeList.stopShimmer()
-                    slEpisodeList.hide()
-                    rvEpisodeList.show()
-                }
-            })
-        }
-
-        private fun observeGetPosts() {
-            activityViewModel.episodeSummaryLiveData.observe(this, Observer {
-                if (it.isEmpty()) {
-                    togglePromptVisibility(true)
-                }
-                else {
-                    fillRecyclerView(it)
-                    togglePromptVisibility(false)
-                }
-            })
-        }
-        private fun togglePromptVisibility(visibility : Boolean){
-            if (visibility) {
-                tvPrompt.visibility = View.VISIBLE
-            } else {
-                tvPrompt.visibility = View.GONE
-            }
-        }
-        private fun fillRecyclerView(list: List<EpisodeSummary>) {
-            Log.d("fill::", "" + list.toString())
-            clearRecyclerView()
-            rvEpisodeList.apply {
-                layoutManager = LinearLayoutManager(context)
-                adapter = MyEpisodeRecyclerViewAdapter(list, this@MainActivity)
-
-            }
-        }
-        private fun clearRecyclerView() {
-            var list = mutableListOf<EpisodeSummary>()
-            list.removeAll(list)
-            rvEpisodeList.apply {
-                layoutManager = LinearLayoutManager(context)
-                adapter = MyEpisodeRecyclerViewAdapter(list, this@MainActivity)
-            }
-        }
-
-        override fun onCreateOptionsMenu(menu: Menu): Boolean {
-            menuInflater.inflate(R.menu.options_menu, menu)
-            val search = menu.findItem(R.id.action_search)
-            val searchView = search.actionView as SearchView
-            searchView.queryHint = getString(R.string.search_hint)
-            searchView.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
-                override fun onQueryTextSubmit(query: String?): Boolean {
-                    if (isOnline()){
-                        lifecycleScope.launch {
-                            activityViewModel.getEpisodesWithLiveData(query)
-                        }
-                        return true
-                    } else
-                        KToasty.warning(this@MainActivity, getString(R.string.no_connection), Toast.LENGTH_SHORT, true).show()
-                    return true
-                }
-                override fun onQueryTextChange(newText: String?): Boolean {
-                    return false
-                }
-            })
-            return true
-        }
-
-        fun isOnline(): Boolean {
-            val connMgr = getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-            val networkInfo: NetworkInfo? = connMgr.activeNetworkInfo
-            return networkInfo?.isConnected == true
-        }
-
-    }*/
 }
