+++ app/src/main/java/com/sevenzeroes/trekkieapp/splash/SplashActivity.kt
@@ -9,8 +9,7 @@ class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        val intent = Intent(this, MainActivity::class.java).apply {
-        }
+        val intent = Intent(this, MainActivity::class.java)
         startActivity(intent)
         finish()
     }
