+++ app/src/main/res/layout/activity_main.xml
@@ -25,12 +25,4 @@
 
 
 
-    <include
-        android:id="@+id/include"
-        layout="@layout/content_main"
-        app:layout_anchor="@+id/button_first"
-        app:layout_anchorGravity="center" />
-
-
-
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
